Reorganises section/task hierarchy
…to ensure that sections are defined strinctly with reference only to a title and an array of tasks, and tasks with a name, slug and a list of pages.

This change also requires the nunjucks directory structure to change, and for tests to be updated.
