+++ server/form-pages/apply/about-the-person/equality-diversity-monitoring/index.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+
+import { Task } from '../../../utils/decorators'
+import EqualityAndDiversity from './equalityAndDiversity'
+
+@Task({
+  name: 'Complete equality and diversity monitoring',
+  slug: 'equality-and-diversity-monitoring',
+  pages: [EqualityAndDiversity],
+})
+export default class EqualityAndDiversityMonitoring {}
