+++ server/form-pages/apply/area-and-funding/funding-information/fundingInformation.ts
@@ -1,8 +1,8 @@
 import type { TaskListErrors } from '@approved-premises/ui'
 import { Cas2Application as Application } from '@approved-premises/api'
-import { Page } from '../../utils/decorators'
-import TaskListPage from '../../taskListPage'
-import { convertKeyValuePairToRadioItems } from '../../../utils/formUtils'
+import { Page } from '../../../utils/decorators'
+import TaskListPage from '../../../taskListPage'
+import { convertKeyValuePairToRadioItems } from '../../../../utils/formUtils'
 
 export const fundingSources = {
   personalSavings: 'Personal money or savings',
