+++ server/form-pages/apply/area-and-funding/funding-information/index.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+
+import { Task } from '../../../utils/decorators'
+import FundingInformationPage from './fundingInformation'
+
+@Task({
+  name: 'Add funding information',
+  slug: 'funding-information',
+  pages: [FundingInformationPage],
+})
+export default class FundingInformation {}
